Solution now console logs the time in hh:mm:ss lcd number format every second.
'0' was added to the number-to-lcd function and there were also minor amends to account for taking the input as a string instead of number (which was necessary due to the addition of the ':' separator).

The time function also now takes into account single digit hr/min/seconds to ensure the time displays properly.